third_party/WebKit/LayoutTests/imported/wpt/html/semantics/document-metadata/the-style-element/style_scoped-expected.txt

This is a testharness.js-based test.
FAIL The scoped attribute is present, the style information must be applied only to its parent element assert_equals: The width should not be applied. expected "50px" but got "100px"
Harness: the test ran to completion.
